Extract Smap content generation into a Module
