+++ test/MemInfo/ProcSpec.hs
@@ -9,9 +9,6 @@ SPDX-License-Identifier: BSD3
 -}
 module MemInfo.ProcSpec (
   spec,
-  genBaseSmap,
-  genSmapLine,
-  genWithPss,
 ) where
 
 import Data.Hashable (hash)
@@ -20,6 +17,7 @@ import Data.Text (Text)
 import qualified Data.Text as Text
 import Data.Word (Word16)
 import Fmt (blockMapF, build, fmt, (+|), (|+))
+import MemInfo.Files.Smap (genBaseSmap, genSmapLine, genWithPss)
 import MemInfo.OrphanInstances ()
 import Numeric.Natural (Natural)
 import System.MemInfo.Proc
@@ -181,13 +179,6 @@ ppZero =
     }
 
 
-genSmapLine :: Text -> Gen (Int, Text)
-genSmapLine prefix = do
-  x <- genValid :: Gen Word16
-  let txt = "" +| prefix |+ ": " +| x |+ " kB"
-  pure (fromIntegral x, txt)
-
-
 genSmap :: Gen (ProcUsage, Text)
 genSmap = oneof [genBaseSmap, genWithSwapPss, genWithPss, genWithPss >>= genAppendSwapPss]
 
@@ -201,51 +192,3 @@ genAppendSwapPss (pp, initial) = do
   (swapPss, txt) <- genSmapLine "SwapPss"
   let content = initial <> "\n" <> txt
   pure (pp {puSwap = swapPss, puMemId = hash content}, content)
-
-
-genWithPss :: Gen (ProcUsage, Text)
-genWithPss = genBaseSmap' >>= genAppendPss
-
-
-genAppendPss :: (ProcUsage, Text, Int) -> Gen (ProcUsage, Text)
-genAppendPss (pp, initial, puPrivateHuge) = do
-  (pss, txt) <- genSmapLine "Pss"
-  let content = initial <> "\n" <> txt
-      newShared = pss - (puPrivate pp - puPrivateHuge)
-  pure (pp {puShared = newShared, puMemId = hash content}, content)
-
-
-genBaseSmap :: Gen (ProcUsage, Text)
-genBaseSmap = do
-  (pp, txt, _) <- genBaseSmap'
-  pure (pp, txt)
-
-
-genBaseSmap' :: Gen (ProcUsage, Text, Int)
-genBaseSmap' = do
-  (clean, cleanTxt) <- genSmapLine "Private_Clean"
-  (dirty, dirtyTxt) <- genSmapLine "Private_Dirty"
-  (sharedClean, shCleanTxt) <- genSmapLine "Shared_Clean"
-  (sharedDirty, shDirtyTxt) <- genSmapLine "Shared_Dirty"
-  (privateHuge, phTxt) <- genSmapLine "Private_Hugetlb"
-  (sharedHuge, shTxt) <- genSmapLine "Shared_Hugetlb"
-  (swap, swapTxt) <- genSmapLine "Swap"
-  let pp =
-        ppZero
-          { puPrivate = clean + dirty + privateHuge
-          , puMemId = hash content
-          , puSwap = swap
-          , puSharedHuge = sharedHuge
-          , puShared = sharedClean + sharedDirty
-          }
-      content =
-        Text.unlines
-          [ phTxt
-          , dirtyTxt
-          , cleanTxt
-          , swapTxt
-          , shTxt
-          , shCleanTxt
-          , shDirtyTxt
-          ]
-  pure (pp, content, privateHuge)
